depreciated registration_colored_icp to use directly registration_icp
