+++ cpp/open3d/pipelines/registration/TransformationEstimation.cpp
@@ -29,6 +29,7 @@
 #include <Eigen/Geometry>
 
 #include "open3d/geometry/PointCloud.h"
+#include "open3d/utility/Console.h"
 #include "open3d/utility/Eigen.h"
 
 namespace open3d {
@@ -92,6 +93,124 @@ Eigen::Matrix4d TransformationEstimationPointToPlane::ComputeTransformation(
         J_r.block<3, 1>(0, 0) = vs.cross(nt);
         J_r.block<3, 1>(3, 0) = nt;
     };
+
+    Eigen::Matrix6d JTJ;
+    Eigen::Vector6d JTr;
+    double r2;
+    std::tie(JTJ, JTr, r2) =
+            utility::ComputeJTJandJTr<Eigen::Matrix6d, Eigen::Vector6d>(
+                    compute_jacobian_and_residual, (int)corres.size());
+
+    bool is_success;
+    Eigen::Matrix4d extrinsic;
+    std::tie(is_success, extrinsic) =
+            utility::SolveJacobianSystemAndObtainExtrinsicMatrix(JTJ, JTr);
+
+    return is_success ? extrinsic : Eigen::Matrix4d::Identity();
+}
+
+double TransformationEstimationForColoredICP::ComputeRMSE(
+        const geometry::PointCloud &source,
+        const geometry::PointCloud &target,
+        const CorrespondenceSet &corres) const {
+    double sqrt_lambda_geometric = sqrt(lambda_geometric_);
+    double lambda_photometric = 1.0 - lambda_geometric_;
+    double sqrt_lambda_photometric = sqrt(lambda_photometric);
+
+    double residual = 0.0;
+    for (size_t i = 0; i < corres.size(); i++) {
+        size_t cs = corres[i][0];
+        size_t ct = corres[i][1];
+        const Eigen::Vector3d &vs = source.points_[cs];
+        const Eigen::Vector3d &vt = target.points_[ct];
+        const Eigen::Vector3d &nt = target.normals_[ct];
+        Eigen::Vector3d vs_proj = vs - (vs - vt).dot(nt) * nt;
+        double is = (source.colors_[cs](0) + source.colors_[cs](1) +
+                     source.colors_[cs](2)) /
+                    3.0;
+        double it = (target.colors_[ct](0) + target.colors_[ct](1) +
+                     target.colors_[ct](2)) /
+                    3.0;
+        const Eigen::Vector3d &dit = target.color_gradient_[ct];
+        double is0_proj = (dit.dot(vs_proj - vt)) + it;
+        double residual_geometric = sqrt_lambda_geometric * (vs - vt).dot(nt);
+        double residual_photometric = sqrt_lambda_photometric * (is - is0_proj);
+        residual += residual_geometric * residual_geometric +
+                    residual_photometric * residual_photometric;
+    }
+    return residual;
+};
+
+Eigen::Matrix4d TransformationEstimationForColoredICP::ComputeTransformation(
+        const geometry::PointCloud &source,
+        const geometry::PointCloud &target,
+        const CorrespondenceSet &corres) const {
+    if (corres.empty()) {
+        utility::LogError(
+                "No correspondences found between source and target "
+                "pointcloud.");
+    }
+    if (!target.HasNormals()) {
+        utility::LogError(
+                "ColoredICP requires target pointcloud to have normals.");
+    }
+    if (!target.HasColors()) {
+        utility::LogError(
+                "ColoredICP requires target pointcloud to have colors.");
+    }
+    if (!source.HasColors()) {
+        utility::LogError(
+                "ColoredICP requires source pointcloud to have colors.");
+    }
+
+    double sqrt_lambda_geometric = sqrt(lambda_geometric_);
+    double lambda_photometric = 1.0 - lambda_geometric_;
+    double sqrt_lambda_photometric = sqrt(lambda_photometric);
+
+    auto compute_jacobian_and_residual =
+            [&](int i,
+                std::vector<Eigen::Vector6d, utility::Vector6d_allocator> &J_r,
+                std::vector<double> &r, std::vector<double> &w) {
+                size_t cs = corres[i][0];
+                size_t ct = corres[i][1];
+                const Eigen::Vector3d &vs = source.points_[cs];
+                const Eigen::Vector3d &vt = target.points_[ct];
+                const Eigen::Vector3d &nt = target.normals_[ct];
+
+                J_r.resize(2);
+                r.resize(2);
+                w.resize(2);
+
+                J_r[0].block<3, 1>(0, 0) = sqrt_lambda_geometric * vs.cross(nt);
+                J_r[0].block<3, 1>(3, 0) = sqrt_lambda_geometric * nt;
+                r[0] = sqrt_lambda_geometric * (vs - vt).dot(nt);
+                w[0] = kernel_->Weight(r[0]);
+
+                // project vs into vt's tangential plane
+                Eigen::Vector3d vs_proj = vs - (vs - vt).dot(nt) * nt;
+                double is = (source.colors_[cs](0) + source.colors_[cs](1) +
+                             source.colors_[cs](2)) /
+                            3.0;
+                double it = (target.colors_[ct](0) + target.colors_[ct](1) +
+                             target.colors_[ct](2)) /
+                            3.0;
+                const Eigen::Vector3d &dit = target.color_gradient_[ct];
+                double is0_proj = (dit.dot(vs_proj - vt)) + it;
+
+                const Eigen::Matrix3d M =
+                        (Eigen::Matrix3d() << 1.0 - nt(0) * nt(0),
+                         -nt(0) * nt(1), -nt(0) * nt(2), -nt(0) * nt(1),
+                         1.0 - nt(1) * nt(1), -nt(1) * nt(2), -nt(0) * nt(2),
+                         -nt(1) * nt(2), 1.0 - nt(2) * nt(2))
+                                .finished();
+
+                const Eigen::Vector3d &ditM = -dit.transpose() * M;
+                J_r[1].block<3, 1>(0, 0) =
+                        sqrt_lambda_photometric * vs.cross(ditM);
+                J_r[1].block<3, 1>(3, 0) = sqrt_lambda_photometric * ditM;
+                r[1] = sqrt_lambda_photometric * (is - is0_proj);
+                w[1] = kernel_->Weight(r[1]);
+            };
 
     Eigen::Matrix6d JTJ;
     Eigen::Vector6d JTr;
